backend/data_ingestion_pipeline/data_ingestion_pipeline/requirements.txt
firecrawl-py
langchain-google-genai
pinecone
sentence-transformers
google-genai
flask
flask-cors
python-dotenv
redis
upstash-redis